+++ CloudySkies/ModApi.cs
@@ -4,6 +4,7 @@ using System.Diagnostics.CodeAnalysis;
 
 using Leclair.Stardew.CloudySkies.Effects;
 using Leclair.Stardew.CloudySkies.LayerData;
+using Leclair.Stardew.CloudySkies.Layers;
 using Leclair.Stardew.CloudySkies.Models;
 using Leclair.Stardew.Common.Types;
 
@@ -27,6 +28,10 @@ public class ModApi : ICloudySkiesApi {
 
 	public void RegenerateLayers(string? weatherId = null) {
 		Mod.UncacheLayers(weatherId);
+	}
+
+	public void RegenerateEffects(string? weatherId = null) {
+		Mod.UncacheEffects(weatherId);
 	}
 
 	public IEnumerable<IWeatherData> GetAllCustomWeather() {
@@ -64,6 +69,9 @@ public class ModApi : ICloudySkiesApi {
 	private static readonly Dictionary<Type, Func<object?>> WeatherDataEditorTypes = new() {
 		{ typeof(IWeatherData), () => new WeatherData() },
 		{ typeof(IScreenTintData), () => new ScreenTintData() },
+
+		{ typeof(ICustomLayerData), () => new CustomLayerData() },
+		{ typeof(ICustomEffectData), () => new CustomEffectData() },
 
 		// When creating our discriminated types, make sure to set Type.
 		{ typeof(IColorLayerData), () => new ColorLayerData() {
@@ -119,4 +127,41 @@ public class ModApi : ICloudySkiesApi {
 			(data, id) => data.Id = id
 		);
 	}
+
+	public bool RegisterLayerType(string type, ICloudySkiesApi.WeatherLayerFactoryDelegate factory) {
+		IWeatherLayer? CreateLayer(ulong id, ILayerData data) {
+			if (data is not ICustomLayerData customData)
+				return null;
+
+			var result = factory(id, customData);
+			return result is null ? null : new WrappedLayer(Mod, result);
+		}
+
+		return Mod.RegisterLayerType(type, CreateLayer);
+	}
+
+	public bool UnregisterLayerType(string type) {
+		return Mod.UnregisterLayerType(type);
+	}
+
+	public bool RegisterEffectType(string type, ICloudySkiesApi.WeatherEffectFactoryDelegate factory) {
+		IWeatherEffect? CreateEffect(ulong id, IEffectData data) {
+			if (data is not ICustomEffectData customData)
+				return null;
+
+			var result = factory(id, customData);
+			return result is null ? null : new WrappedEffect(Mod, result);
+		}
+
+		return Mod.RegisterEffectType(type, CreateEffect);
+	}
+
+	public bool UnregisterEffectType(string type) {
+		return Mod.UnregisterEffectType(type);
+	}
+
+	public void NotifyLoadsAsset(ulong id, string assetName) {
+		Mod.MarkLoadsAsset(id, assetName);
+	}
+
 }