+++ Common/Serialization/Converters/DiscriminatingConverter.cs
@@ -23,10 +23,12 @@ public class DiscriminatedType : Attribute {
 public class DiscriminatingConverter<T> : JsonConverter where T : class {
 
 	public readonly string Key;
+	public readonly string? WildcardKey;
 	private readonly Dictionary<string, Type> Types;
 
-	public DiscriminatingConverter(string key, Dictionary<string, Type> types) {
+	public DiscriminatingConverter(string key, Dictionary<string, Type> types, string? wildcardKey = null) {
 		Key = key;
+		WildcardKey = wildcardKey;
 		Types = types;
 	}
 
@@ -57,16 +59,14 @@ public class DiscriminatingConverter<T> : JsonConverter where T : class {
 		if (string.IsNullOrEmpty(type))
 			throw new JsonReaderException($"Cannot read \"{Key}\" from {reader.TokenType} node (path: {reader.Path}).");
 
-		if (!Types.TryGetValue(type, out Type? value))
+		if (!Types.TryGetValue(type, out Type? value) && (WildcardKey == null || !Types.TryGetValue(WildcardKey, out value)))
 			throw new JsonReaderException($"Invalid type \"{type}\" read from \"{Key}\" of {reader.TokenType} node (path: {reader.Path}).");
 
-		object? result = Activator.CreateInstance(value);
-		if (result is null)
-			throw new JsonReaderException($"Unable to create new instance of type {value.FullName ?? value.Name} (type: \"{type}\", path: {reader.Path}).");
+		object? result = Activator.CreateInstance(value)
+			?? throw new JsonReaderException($"Unable to create new instance of type {value.FullName ?? value.Name} (type: \"{type}\", path: {reader.Path}).");
 
 		serializer.Populate(obj.CreateReader(), result);
 		return result;
-
 	}
 
 	public override bool CanWrite => false;